authorization-for-rag/openai-fga-python/requirements.txt
faiss-cpu==1.9.0.post1
numpy==2.2.2
openai==1.59.7
openfga-sdk==0.9.0